Move use case => lib/
